==> 08_capas_complementarias.ipynb <==
import os
from pathlib import Path

import geopandas as gpd
import shapely

from nad.load import load_ageb_geometry

data_path = Path(os.environ["DATA_PATH"])
population_grids_path = Path(os.environ["POPULATION_GRIDS_PATH"])

results_path = Path("./results")
results_path.mkdir(exist_ok=True, parents=True)
(results_path / "platform").mkdir(exist_ok=True, parents=True)

# Limite urbano

ageb_geom = load_ageb_geometry(
    population_grids_path / "final" / "zone_agebs" / "shaped" / "2020" / "08.2.03.gpkg",
)

df_neighborhoods = (gpd.read_file(data_path / "datos" / "colonias_raw.geojson")).to_crs(
    ageb_geom.crs,
)


df_limit = gpd.GeoDataFrame(
    geometry=[
        shapely.union(
            ageb_geom.union_all(),
            df_neighborhoods["geometry"].union_all(),
        ).buffer(10),
    ],
    crs=ageb_geom.crs,
).to_crs("EPSG:4326")

df_limit.to_file(results_path / "platform" / "limite_urbano.geojson")

# Vialidades

df_roads = gpd.read_file(data_path / "datos" / "SAP_Vialildad_Primaria.geojson")
df_roads.to_crs("EPSG:4326").to_file(results_path / "platform" / "vialidades.geojson")
