==> 11_arboles.ipynb <==
import os
from pathlib import Path

import ee
import geemap
import numpy as np
import shapely

from nad.load import load_equip_df

ee.Initialize(project="ee-ursa-test")

data_path = Path(os.environ["DATA_PATH"])

df = (
    load_equip_df(
        data_path / "datos" / "pparquesWgs84",
        extra_cols=["CLASIFICAC"],
    )
    .assign(geometry=lambda df: df["geometry"].make_valid())
    .to_crs("EPSG:4326")
)

df = df[df["geometry"].geom_type == "Polygon"].reset_index(names="polygon_id")

bounds = ee.geometry.Geometry.Polygon(
    list(shapely.box(*df.total_bounds).exterior.coords),
)

chunk_size = 4000
df_chunks = np.array_split(df, 2)

feature_chunks = [geemap.geopandas_to_ee(chunk) for chunk in df_chunks]

img_canopy = (
    ee.ImageCollection("projects/sat-io/open-datasets/facebook/meta-canopy-height")
    .filterBounds(bounds)
    .first()
)

mask = img_canopy.gte(3).unmask(0)

out = ee.data.computeFeatures(
    {
        "expression": ee.image.Image.pixelArea()
        .multiply(mask)
        .reduceRegions(
            feature_chunks[0],
            reducer=ee.reducer.Reducer.sum(),
            scale=1,
            crs="EPSG:4326",
        ),
        "fileFormat": "GEOPANDAS_GEODATAFRAME",
    },
)

out = (
    out.set_crs("EPSG:4326")
    .to_crs("EPSG:6372")
    .assign(
        area_frac=lambda df: df["sum"] / df["geometry"].area.sum(),
    )
    .rename(columns={"sum": "area"})
)

out.query("area == 0")["CLASIFICAC"].value_counts()
